financeservice/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-dotenv>=1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
aiohttp>=3.8.5
uvloop==0.19.0
httptools==0.6.1
email_validator
passlib[bcrypt]==1.7.4
shortuuid==1.0.13
python-jose[cryptography]
redis==5.2.1
APScheduler>=3.10.4 
supabase==1.1.0